Split library into smaller files
