+++ src/main/kotlin/net/emdal/tank/Entity.kt
@@ -1,0 +1,3 @@
+package net.emdal.tank
+
+open class Entity
