+++ src/main/kotlin/net/emdal/tank/Graph.kt
@@ -1,0 +1,3 @@
+package net.emdal.tank
+
+data class Graph(val query: String = "", val aliases: List<Pair<String, Entity>> = emptyList())
