+++ src/main/kotlin/net/emdal/tank/Query.kt
@@ -1,0 +1,10 @@
+package net.emdal.tank
+
+import org.neo4j.driver.Session
+
+fun Graph.match(graph: Graph.() -> Graph): Graph {
+  return this.copy(query = "MATCH ").graph()
+}
+
+fun Session.query(block: Graph.() -> Graph) =
+  this.run("${Graph().block().query} RETURN *").list().toList()
